embed-able, fixed test
