+++ example/foo.md
@@ -1,1 +1,1 @@
-_Updated on <!--tmpl,chomp:date-->Sun 31 Mar 2019 00:37:25 AEDT<!--/tmpl-->_
+_Updated on <!--tmpl,chomp:date -->Wed 10 Apr 14:37:06 AEST 2019<!--/tmpl-->_
