+++ example/html.md
@@ -1,1 +1,1 @@
-<!--tmpl:echo -n This is a \<a href=\"foo\"\-->This is a <a href="foo"<!--/tmpl-->
+<!--tmpl:echo -n This is a \<a href=\"foo\"\ -->This is a <a href="foo"<!--/tmpl-->
