+++ go.mod
@@ -1,0 +1,7 @@
+module github.com/jpillora/md-tmpl
+
+go 1.12
+
+require github.com/jpillora/opts v0.0.0-20160806153215-3f7962811f23
+
+replace github.com/jpillora/opts => ../millergarym-opts/
